Cleanup and refactor code
