+++ krystex/src/main/java/com/flipkart/krystal/krystex/commands/ExecuteWithDependency.java
@@ -1,8 +1,8 @@
 package com.flipkart.krystal.krystex.commands;
 
 import com.flipkart.krystal.data.Results;
-import com.flipkart.krystal.krystex.RequestId;
 import com.flipkart.krystal.krystex.node.NodeId;
+import com.flipkart.krystal.krystex.request.RequestId;
 
 public record ExecuteWithDependency(
     NodeId nodeId, String dependencyName, Results<Object> results, RequestId requestId)
