+++ krystex/src/main/java/com/flipkart/krystal/krystex/commands/ExecuteWithInputs.java
@@ -1,9 +1,9 @@
 package com.flipkart.krystal.krystex.commands;
 
 import com.flipkart.krystal.data.Inputs;
-import com.flipkart.krystal.krystex.RequestId;
 import com.flipkart.krystal.krystex.node.DependantChain;
 import com.flipkart.krystal.krystex.node.NodeId;
+import com.flipkart.krystal.krystex.request.RequestId;
 import com.google.common.collect.ImmutableSet;
 
 public record ExecuteWithInputs(
