+++ krystex/src/main/java/com/flipkart/krystal/krystex/commands/NodeRequestCommand.java
@@ -1,6 +1,6 @@
 package com.flipkart.krystal.krystex.commands;
 
-import com.flipkart.krystal.krystex.RequestId;
+import com.flipkart.krystal.krystex.request.RequestId;
 
 public sealed interface NodeRequestCommand extends NodeCommand
     permits ExecuteWithDependency, ExecuteWithInputs, SkipNode {
