+++ krystex/src/main/java/com/flipkart/krystal/krystex/commands/SkipNode.java
@@ -1,9 +1,9 @@
 package com.flipkart.krystal.krystex.commands;
 
-import com.flipkart.krystal.krystex.RequestId;
 import com.flipkart.krystal.krystex.ResolverCommand.SkipDependency;
 import com.flipkart.krystal.krystex.node.DependantChain;
 import com.flipkart.krystal.krystex.node.NodeId;
+import com.flipkart.krystal.krystex.request.RequestId;
 
 public record SkipNode(
     NodeId nodeId,
